Cleanup TrackedInputContext exports
